+++ src/meta/api/src/lib.rs
@@ -24,6 +24,7 @@ mod data_mask_api_impl;
 mod database_api;
 pub mod database_util;
 pub mod dictionary_api;
+pub mod garbage_collection_api;
 pub mod index_api;
 pub mod kv_app_error;
 pub mod kv_pb_api;
@@ -53,6 +54,7 @@ pub use catalog_api::CatalogApi;
 pub use data_mask_api::DatamaskApi;
 pub use database_api::DatabaseApi;
 pub use dictionary_api::DictionaryApi;
+pub use garbage_collection_api::GarbageCollectionApi;
 pub use index_api::IndexApi;
 pub use lock_api::LockApi;
 pub use row_access_policy_api::RowAccessPolicyApi;
