+++ src/meta/api/src/table_api.rs
@@ -111,6 +111,8 @@ use crate::database_api::DatabaseApi;
 use crate::database_util::get_db_or_err;
 use crate::deserialize_struct;
 use crate::fetch_id;
+use crate::garbage_collection_api::get_history_tables_for_gc;
+use crate::garbage_collection_api::ORPHAN_POSTFIX;
 use crate::get_u64_value;
 use crate::kv_app_error::KVAppError;
 use crate::kv_pb_api::KVPbApi;
@@ -121,10 +123,8 @@ use crate::schema_api::construct_drop_table_txn_operations;
 use crate::schema_api::get_db_by_id_or_err;
 use crate::schema_api::get_db_id_or_err;
 use crate::schema_api::get_history_table_metas;
-use crate::schema_api::get_history_tables_for_gc;
 use crate::schema_api::handle_undrop_table;
 use crate::schema_api::table_has_to_not_exist;
-use crate::schema_api::ORPHAN_POSTFIX;
 use crate::send_txn;
 use crate::serialize_struct;
 use crate::serialize_u64;
